pyproject.toml
[tool.poetry]
name = "ide-proxy"
version = "0.1.0"
description = ""
authors = ["Tarun Luthra <tarunluthra987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
django-rest-framework = "^0.1.0"
black = "^21.6b0"
python-dotenv = "^0.18.0"
requests = "^2.25.1"
django-cors-headers = "^3.7.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
